Resolve merge conflicts
