+++ examples/cpp/local/CMakeLists.txt
@@ -9,7 +9,7 @@ Include(FetchContent)
 FetchContent_Declare(
   Catch2
   GIT_REPOSITORY https://github.com/catchorg/Catch2.git
-  GIT_TAG        v3.4.0 # or a later release
+  GIT_TAG        v3.5.0 # or a later release
 )
 FetchContent_Declare(
   cpprealm
